+++ tests/wdio/loaders-spec.js
@@ -1,26 +1,26 @@
 Terra.describeViewports('loaders', ['huge'], () => {
   it('loads a js codeblock', () => {
-    browser.url('/raw/test/terra-dev-site/loaders/codeblock-js');
+    browser.url('/raw/test/cerner-terra-dev-site/loaders/codeblock-js');
     Terra.validates.element('js codeblock');
   });
 
   it('loads a json codeblock', () => {
-    browser.url('/raw/test/terra-dev-site/loaders/codeblock-json');
+    browser.url('/raw/test/cerner-terra-dev-site/loaders/codeblock-json');
     Terra.validates.element('json codeblock');
   });
 
   it('loads a ts codeblock', () => {
-    browser.url('/raw/test/terra-dev-site/loaders/codeblock-ts');
+    browser.url('/raw/test/cerner-terra-dev-site/loaders/codeblock-ts');
     Terra.validates.element('ts codeblock');
   });
 
   it('loads a scss codeblock', () => {
-    browser.url('/raw/test/terra-dev-site/loaders/codeblock-scss');
+    browser.url('/raw/test/cerner-terra-dev-site/loaders/codeblock-scss');
     Terra.validates.element('scss codeblock');
   });
 
   it('loads a package', () => {
-    browser.url('/raw/test/terra-dev-site/loaders/package');
+    browser.url('/raw/test/cerner-terra-dev-site/loaders/package');
     Terra.validates.element('package');
   });
 
@@ -28,7 +28,7 @@ Terra.describeViewports('loaders', ['huge'], () => {
     const viewport = browser.getViewportSize();
     viewport.height = 3000;
     browser.setViewportSize(viewport);
-    browser.url('/raw/test/terra-dev-site/loaders/props-table');
+    browser.url('/raw/test/cerner-terra-dev-site/loaders/props-table');
     Terra.validates.element('props table');
   });
 
@@ -36,7 +36,7 @@ Terra.describeViewports('loaders', ['huge'], () => {
     const viewport = browser.getViewportSize();
     viewport.height = 3000;
     browser.setViewportSize(viewport);
-    browser.url('/raw/test/terra-dev-site/loaders/example');
+    browser.url('/raw/test/cerner-terra-dev-site/loaders/example');
     Terra.validates.element('example');
   });